+++ public/controllers/agent/components/agents-preview.js
@@ -120,7 +120,7 @@ export const AgentsPreview = compose(
         this.setState({ platforms: platformsModel, loading: false });
     } catch (error) {}
   }
-  
+
   removeFilters(){
     this._isMount && this.setState({agentTableFilters: []})
   }
@@ -164,7 +164,7 @@ export const AgentsPreview = compose(
             </EuiFlexItem>
             {this.totalAgents > 0 && (
               <EuiFlexItem >
-                <EuiPanel betaBadgeLabel="Details">  
+                <EuiPanel betaBadgeLabel="Details">
                   <EuiFlexGroup>
                     <EuiFlexItem>
                       {this.summary && (
@@ -236,7 +236,7 @@ export const AgentsPreview = compose(
                               <EuiToolTip
                                 position='top'
                                 content='View agent details'>
-                                  <a onClick={() => 
+                                  <a onClick={() =>
                                   this.props.tableProps.showAgent(
                                     this.lastAgent
                                   )}>{this.lastAgent.name}</a>
@@ -271,7 +271,7 @@ export const AgentsPreview = compose(
                               style={{ whiteSpace: 'nowrap' }}
                               titleSize="s"
                               description="Most active agent"
-                              titleColor="primary"                              
+                              titleColor="primary"
                             />
                           </EuiFlexItem>
                         )}
@@ -298,13 +298,13 @@ export const AgentsPreview = compose(
                     </div>
                     {this.props.resultState === 'loading' &&
                       (
-                      <div style={{ display: 'block', textAlign: "center", padding: 30}}>                        
+                      <div style={{ display: 'block', textAlign: "center", padding: 30}}>
                         <EuiLoadingChart size="xl" />
                       </div>
                     ) }
-                      
+
                     </EuiFlexItem>
-                  </EuiFlexGroup>                  
+                  </EuiFlexGroup>
                 </EuiPanel>
                 <EuiPanel paddingSize="none" betaBadgeLabel="Evolution" style={{ height: 180,  display: this.props.resultState === 'none' ? 'block' : 'none'}}>
                   <EuiEmptyPrompt
@@ -318,7 +318,7 @@ export const AgentsPreview = compose(
                   />
                 </EuiPanel>
               </EuiFlexItem>
-              
+
             )}
           </EuiFlexGroup>
           <EuiSpacer size="m" />